% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myconstr.R
\name{myconstr}
\alias{myconstr}
\title{A constructor function for t tests}
\usage{
myconstr(x, y, alpha = 0.05)
}
\arguments{
\item{x}{#vector of data}

\item{y}{#vector of data}

\item{alpha}{#the alpha, a value between 0-1}
}
\value{
A named list with a dataframe of the data, confidence interval, and p-value
}
\description{
This function takes in two vectors of data and an alpha value, and
it provides the confidence interval and p-value from a standard t-test.
NOTE: as the function is now, it assumes the variances of the two vectors of data are the same.
}
\examples{
x <- rnorm(30,5,2); y <- rnorm(30,3,2)
}
