% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{myttest}
\alias{myttest}
\title{A constructor function for t tests}
\usage{
myttest(x, y, paired = FALSE, alpha = 0.05)
}
\arguments{
\item{x}{#vector of data}

\item{y}{#vector of data}

\item{paired}{#a boolean that shows whether the supplied data is paired. This value defaults to false.}

\item{alpha}{#the alpha, a value between 0-1}
}
\value{
A named list with a dataframe of the data, confidence interval, and p-value
}
\description{
This function takes in two vectors of data, true/false of whether the data is paired, and an alpha value for the acceptable error
It provides the confidence interval and p-value from a t-test.
}
\examples{
x <- rnorm(30,5,2); y <- rnorm(30,3,2)
obj <- myttest(x=x, y=y, paired=FALSE, alpha=0.05)
}
