% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myconstr.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{A plot method for Rttest objects}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{#An object of class Rttest. This is a list with data, confidence interval, and pvalue}

\item{...}{#any additional arguments to the plot function.}
}
\value{
A formatted table of the input data with a note at the bottom with the confidence interval and the p-value
}
\description{
This function plots the data that was supplied when creating a Rttest object. The plot is of two boxplots.
}
