% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myconstr.R
\name{print.Rttest}
\alias{print.Rttest}
\title{A constructor function for t tests}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{#An object of class Rttest. This is a list with data, confidence interval, and pvalue}

\item{...}{#allows the function to work if more than one argument is submitted.}
}
\value{
A formatted table of the input data
}
\description{
This function prints a Rttest object.
}
