% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{A plot method for Rttest objects}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{#An object of class Rttest. This is a list.}

\item{...}{#any additional arguments to the plot function.}
}
\value{
Side-by-side boxplots that show the data distribution of the supplied populations or a single boxplot of the difference of the data, depending on whether the data is paired.
}
\description{
This function plots the data that was supplied when creating a Rttest object, if the data is non-paired. The plot is of two boxplots. However, if the data is paired, one boxplot of the difference of data is shown.
}
