% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{A plot method for Rttest objects}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{#An object of class Rttest. This is a list with data, confidence interval, and pvalue}

\item{...}{#any additional arguments to the plot function.}
}
\value{
Side-by-side boxplots that show the data distribution of all supplied populations.
}
\description{
This function plots the data that was supplied when creating a Rttest object. The plot is of two boxplots.
}
