% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{print.Rttest}
\alias{print.Rttest}
\title{A print method for Rttest objects}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{#An object of class Rttest. This is a list with data, confidence interval, and pvalue}

\item{...}{#allows the function to work if more than one argument is submitted.}
}
\value{
A string that has the confidence interval for the provided alpha.
}
\description{
This function prints the confidence interval for an Rttest object.
}
